+++ tests/testthat/test-crop.R
@@ -4,4 +4,10 @@ test_that("AFM image cropping", {
   q = summary(d1)
   expect_equal(q$resolution, "50 x 50")
   expect_equal(q$size, "490 x 490 nm")
+
+  # try cropping beyond margins  
+  d1 = AFM.crop(d, x0=85, y0=45, width.pixels = 50, height.pixels = 50)
+  q = summary(d1)
+  expect_equal(q$resolution, "15 x 50")
+  expect_equal(q$size, "140 x 490 nm")
 })